Rename packages to compliance
