+++ src/main/java/constants/ErrorMessages.java
@@ -1,4 +1,4 @@
-package Constants;
+package constants;
 
 public class ErrorMessages {
     public static final String CREATE_USAGE = "Usage: add q/{QUESTION} a/{ANSWER}";
