+++ src/main/java/parser/Parser.java
@@ -1,9 +1,9 @@
-package Parser;
+package parser;
 
-import static Constants.CommandConstants.CREATE;
+import static constants.CommandConstants.CREATE;
 
-import Commands.Command;
-import Commands.CommandCreate;
+import command.Command;
+import command.CommandCreate;
 
 public class Parser {
     public static Command parseInput(String input) throws IllegalArgumentException {
@@ -11,8 +11,8 @@ public class Parser {
         String command = inputList[0];
         String arguments = inputList.length > 1 ? inputList[1] : "";
         return switch (command) {
-            case CREATE -> new CommandCreate(arguments);
-            default -> throw new IllegalArgumentException();
+        case CREATE -> new CommandCreate(arguments);
+        default -> throw new IllegalArgumentException();
         };
     }
 }
